+++ app/src/main/java/it/gym/facade/ReservationFacade.java
@@ -70,7 +70,7 @@ public class ReservationFacade {
 
         isReservedOnTime(startTime);
 
-        List<AEvent> timesOff = this.eventService.findAllEventsTypeInBetween(startTime, endTime);
+        List<AEvent> timesOff = this.eventService.findAllEventsLargerThanInterval(startTime, endTime);
 
         hasHolidays(timesOff);